c1

+++ WebVella.Erp.Plugins.Mail/MailPlugin.20190419.cs
@@ -300,7 +300,7 @@ namespace WebVella.Erp.Plugins.Mail
 				var url = "/sdk/server/job/l/plan";
 				var iconClass = "fa fa-cogs";
 				var weight = 1;
-				var type = ((int)3);
+				var type = 3;
 				var access = new List<Guid>();
 				var entityListPages = new List<Guid>();
 				var entityCreatePages = new List<Guid>();
